+++ examples/bootstrap.py
@@ -3,7 +3,7 @@
 #
 #The MIT License (MIT)
 #
-#Copyright (c) <2013> <Colin Duquesnoy and others, see AUTHORS.txt>
+#Copyright (c) <2013-2014> <Colin Duquesnoy and others, see AUTHORS.txt>
 #
 #Permission is hereby granted, free of charge, to any person obtaining a copy
 #of this software and associated documentation files (the "Software"), to deal
@@ -29,6 +29,7 @@ show a splash screen while preloading.
 """
 import sys
 import logging
+logging.basicConfig(level=logging.INFO)
 import pyqode.core
 import pyqode.python
 from pyqode.qt import QtGui
@@ -60,5 +61,4 @@ def main():
 
 
 if __name__ == "__main__":
-    logging.basicConfig(level=logging.DEBUG)
     main()